+++ src/main/java/me/shadowalzazel/mcodyssey/phenomenons/DailyPhenomena.kt
@@ -1,0 +1,25 @@
+package me.shadowalzazel.mcodyssey.phenomenons
+
+import me.shadowalzazel.mcodyssey.phenomenons.dailyPhenomena.*
+import me.shadowalzazel.mcodyssey.phenomenons.dailyPhenomena.utilty.DailyPhenomenon
+
+object DailyPhenomena {
+
+    val BLAZING_SOUL: DailyPhenomenon = BlazingSouls
+    val BREEZY_WINDS: DailyPhenomenon = BreezyWinds
+    val COMET_SIGHTING: DailyPhenomenon = CometSighting
+    val DANCE_OF_THE_BIOLUMINESCENT: DailyPhenomenon = DanceOfTheBioluminescent
+    val EARTHQUAKE: DailyPhenomenon = Earthquake
+    val FAMINE: DailyPhenomenon = Famine
+    val FLAUNTING_FAIRIES: DailyPhenomenon = FlauntingFairies
+    val GRAVITY_SHIFT: DailyPhenomenon = GravityShift
+    val SHIMMER_INTOXICATION: DailyPhenomenon = ShimmerIntoxication
+    val SLIME_STORM: DailyPhenomenon = SlimeStorm
+    val SOLAR_FLARE: DailyPhenomenon = SolarFlare
+    val SPIRITS_RELEASE: DailyPhenomenon = SpiritsRelease
+    val STONE_FLASH: DailyPhenomenon = StoneFlash
+    //
+    val phenomenaList = listOf(BLAZING_SOUL, BREEZY_WINDS, COMET_SIGHTING, DANCE_OF_THE_BIOLUMINESCENT, EARTHQUAKE, FAMINE, FLAUNTING_FAIRIES, GRAVITY_SHIFT, SHIMMER_INTOXICATION, SLIME_STORM, SOLAR_FLARE,
+            SPIRITS_RELEASE, STONE_FLASH)
+
+}
