+++ src/main/java/me/shadowalzazel/mcodyssey/phenomenons/NightlyPhenomena.kt
@@ -1,0 +1,13 @@
+package me.shadowalzazel.mcodyssey.phenomenons
+
+import me.shadowalzazel.mcodyssey.phenomenons.nightlyPhenomena.*
+import me.shadowalzazel.mcodyssey.phenomenons.nightlyPhenomena.utilty.NightlyPhenomenon
+
+object NightlyPhenomena {
+
+    val BLOOD_MOON: NightlyPhenomenon = BloodMoon
+    val BLUE_MOON: NightlyPhenomenon = BlueMoon
+    //
+    val phenomenaList = listOf(BLOOD_MOON, BLUE_MOON)
+
+}
