+++ src/main/java/me/shadowalzazel/mcodyssey/phenomenons/dailyPhenomena/SlimeStorm.kt
@@ -1,0 +1,13 @@
+package me.shadowalzazel.mcodyssey.phenomenons.dailyPhenomena
+
+import me.shadowalzazel.mcodyssey.phenomenons.dailyPhenomena.utilty.DailyPhenomenon
+import org.bukkit.World
+
+object SlimeStorm : DailyPhenomenon("Slime_Storm", 80, 5, 50) {
+
+    override fun phenomenonEffect(phenomenonWorld: World){
+        println("Slime is falling from the Sky at ${phenomenonWorld.name}!")
+
+    }
+
+}
