--- src/main/java/me/shadowalzazel/mcodyssey/phenomenons/megaPhenomenon.kt
@@ -1,6 +1,0 @@
-package me.shadowalzazel.mcodyssey.phenomenons
-
-import me.shadowalzazel.mcodyssey.MinecraftOdyssey
-import org.bukkit.ChatColor
-import org.bukkit.World
-
