--- src/main/java/me/shadowalzazel/mcodyssey/phenomenons/newAdvancements.kt
@@ -1,34 +1,0 @@
-package me.shadowalzazel.mcodyssey.phenomenons
-
-import io.papermc.paper.advancement.AdvancementDisplay
-import org.bukkit.NamespacedKey
-import org.bukkit.advancement.Advancement
-
-object Pentakill: Advancement {
-    override fun getKey(): NamespacedKey {
-        TODO("Not yet implemented")
-    }
-    override fun getCriteria(): MutableCollection<String> {
-
-        TODO("Not yet implemented")
-    }
-
-    override fun getDisplay(): AdvancementDisplay? {
-        TODO("Not yet implemented")
-    }
-
-    override fun getParent(): Advancement? {
-        TODO("Not yet implemented")
-    }
-
-    override fun getChildren(): MutableCollection<Advancement> {
-        TODO("Not yet implemented")
-    }
-
-    override fun getRoot(): Advancement {
-        TODO("Not yet implemented")
-    }
-
-}
-
-
