+++ src/main/java/me/shadowalzazel/mcodyssey/tasks/misc.kt
@@ -1,0 +1,2 @@
+package me.shadowalzazel.mcodyssey.tasks
+
